chore: add cta for no created tables in /auth/policies
